try to fix clang compilation complains.

+++ lib/ccv_util.c
@@ -522,7 +522,7 @@ int ccv_matrix_eq(ccv_matrix_t* a, ccv_matrix_t* b)
 		{ \
 			for (j = 0; j < da->cols * ch; j++) \
 			{ \
-				if (fabs(_for_get(b_ptr, j, 0) - _for_get(a_ptr, j, 0)) > epsilon) \
+				if (fabs((double)(_for_get(b_ptr, j, 0) - _for_get(a_ptr, j, 0))) > epsilon) \
 					return -1; \
 			} \
 			a_ptr += da->step; \
